# docs/openapi/platform-openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Observability Platform API",
    "description": "File n�y du?c g?p th? c�ng t? /v3/api-docs c?a ingestion-service v� core-app t?i th?i di?m t?o � c� th? l?ch n?u code sau n�y d?i m� kh�ng regenerate l?i; ngu?n s? th?t lu�n l� 2 endpoint /v3/api-docs s?ng.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/services": {
      "get": {
        "tags": [
          "Service Management"
        ],
        "summary": "List all services",
        "description": "Returns a list of all registered services. Yêu cầu role: ADMIN, DEVOPS.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Successfully retrieved list of services",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseListServiceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing ADMIN or DEVOPS role)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Service Management"
        ],
        "summary": "Register a new service",
        "description": "Registers a new service and returns a plaintext API Key. WARNING: The plaintext API Key is returned only once in this response. Yêu cầu role: ADMIN.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Service registered successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseServiceCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing ADMIN role)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "409": {
            "description": "Service already exists",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/services/{id}/status": {
      "patch": {
        "tags": [
          "Service Management"
        ],
        "summary": "Update service status",
        "description": "Updates the status of a specific service. Yêu cầu role: ADMIN.",
        "operationId": "updateStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateServiceStatusRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Status updated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseServiceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden (missing ADMIN role)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Service not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/logs": {
      "get": {
        "tags": [
          "Log Search"
        ],
        "summary": "Search logs",
        "description": "Search logs with various filters. Note: Pagination can be done via `page`/`size` (up to 10k limit) OR via `search_after` for deep pagination. When searching by `traceId`, results are sorted ascending by timestamp. Yêu cầu role: ADMIN, DEVOPS, DEVELOPER, VIEWER (authenticated).",
        "operationId": "searchLogs",
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "description": "Filter by service name",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "description": "Filter by environment",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "in": "query",
            "description": "Filter by log level",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "Start of time range (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "End of time range (ISO-8601)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "traceId",
            "in": "query",
            "description": "Filter by trace ID (forces ASC sorting)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "spanId",
            "in": "query",
            "description": "Filter by span ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "host",
            "in": "query",
            "description": "Filter by host",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statusCode",
            "in": "query",
            "description": "Filter by HTTP status code",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "httpMethod",
            "in": "query",
            "description": "Filter by HTTP method",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endpoint",
            "in": "query",
            "description": "Filter by API endpoint",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "description": "Full text search query on log message",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (0-based) for standard pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Page size for standard pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "search_after",
            "in": "query",
            "description": "Cursor for deep pagination. Use the `nextSearchAfter` value from the previous response.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search successful",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponseLogSearchResponseObject"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error (e.g. Elasticsearch failure)",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UnifiedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UnifiedResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "object",
            "description": "The main response data payload"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "CreateServiceRequest": {
        "required": [
          "environment",
          "name",
          "teamOwner"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "teamOwner": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "environment": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ServiceCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "plaintextApiKey": {
            "type": "string"
          },
          "keyPrefix": {
            "type": "string"
          }
        },
        "description": "The main response data payload"
      },
      "UnifiedResponseServiceCreateResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "$ref": "#/components/schemas/ServiceCreateResponse"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "UpdateServiceStatusRequest": {
        "required": [
          "status"
        ],
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "DISABLED"
            ]
          }
        }
      },
      "ServiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "teamOwner": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "description": "The main response data payload"
      },
      "UnifiedResponseServiceResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "$ref": "#/components/schemas/ServiceResponse"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
      "UnifiedResponseListServiceResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "type": "array",
            "description": "The main response data payload",
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            }
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
      "LogSearchResponseObject": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "description": "List of matching log events",
            "items": {
              "type": "object",
              "description": "List of matching log events"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "description": "Total number of elements matching the query",
            "format": "int64",
            "example": 1500
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "nextSearchAfter": {
            "type": "string",
            "description": "Cursor value for deep pagination. Pass this as `search_after` in the next request.",
            "example": "1694167530000,12345"
          }
        },
        "description": "Paginated response for log search results"
      },
      "UnifiedResponseLogSearchResponseObject": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Status code of the response (e.g., SUCCESS, VALIDATION_ERROR, NOT_FOUND)",
            "example": "SUCCESS"
          },
          "data": {
            "$ref": "#/components/schemas/LogSearchResponseObject"
          },
          "message": {
            "type": "string",
            "description": "Message providing more context, usually on errors"
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response (ISO-8601)",
            "example": "2026-09-08T10:15:30Z"
          },
          "path": {
            "type": "string",
            "description": "API path that was called",
            "example": "/api/v1/services"
          },
          "fieldErrors": {
            "type": "array",
            "description": "List of validation errors if status is VALIDATION_ERROR",
            "items": {
              "type": "object",
              "additionalProperties": {
                "type": "string",
                "description": "List of validation errors if status is VALIDATION_ERROR"
              },
              "description": "List of validation errors if status is VALIDATION_ERROR"
            }
          }
        },
        "description": "Unified envelope response format for all API calls"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "X-API-Key": {
        "type": "apiKey",
        "description": "API Key authentication. Required for all endpoints.",
        "name": "X-API-Key",
        "in": "header"
      }
    }
  }
}